Using pad_to_size function to remove duplicate code in pad_arrays (#1047)
* Using pad_to_size function to remove deuplicate code in pad_arrays and moving to util

* Removed test_pad_arrays_axis from test_variable_length

* Moving pad_arrays back to mx.batchify since it uses mxnet

* Added unused multi_processing test param back to have the same number of tests so that the flaky tests pass

* Fixing black
