+++ src/gluonts/mx/batchify.py
@@ -19,6 +19,7 @@ import mxnet as mx
 
 from gluonts.core.component import DType
 from gluonts.dataset.common import DataBatch
+from gluonts.support.util import pad_to_size
 
 
 # TODO: should the following contempate mxnet arrays or just numpy arrays?
@@ -35,7 +36,7 @@ def _is_stackable(arrays: List, axis: int = 0) -> bool:
 
 # TODO: should the following contempate mxnet arrays or just numpy arrays?
 def _pad_arrays(
-    data: List[Union[np.ndarray, mx.nd.NDArray]], axis: int = 0,
+    data: List[Union[np.ndarray, mx.nd.NDArray]], axis: int = 0
 ) -> List[Union[np.ndarray, mx.nd.NDArray]]:
     assert isinstance(data[0], (np.ndarray, mx.nd.NDArray))
     is_mx = isinstance(data[0], mx.nd.NDArray)
@@ -51,12 +52,9 @@ def _pad_arrays(
         if is_mx:
             x = x.asnumpy()
 
-        pad_size = max_len - x.shape[axis]
-        pad_lengths = [(0, 0)] * x.ndim
-        pad_lengths[axis] = (0, pad_size)
-        x_padded = np.pad(x, mode="constant", pad_width=pad_lengths)
+        x = pad_to_size(x, max_len, axis)
 
-        padded_data.append(x_padded if not is_mx else mx.nd.array(x_padded))
+        padded_data.append(x if not is_mx else mx.nd.array(x))
 
     return padded_data
